Refactoring Code base and Debugging Audio
